Refactor EholdingsPackagesImpl to use RMAPITemplate
