+++ src/main/java/org/folio/rest/converter/PackageByIdConverter.java
@@ -1,0 +1,19 @@
+package org.folio.rest.converter;
+
+import org.folio.rest.jaxrs.model.Package;
+import org.folio.rmapi.model.PackageByIdData;
+import org.folio.rmapi.result.PackageResult;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.convert.converter.Converter;
+import org.springframework.stereotype.Component;
+
+@Component
+public class PackageByIdConverter implements Converter<PackageByIdData, Package> {
+  @Autowired
+  private Converter<PackageResult, Package> packageConverter;
+
+  @Override
+  public Package convert(PackageByIdData packageByIdData) {
+    return packageConverter.convert(new PackageResult(packageByIdData, null, null));
+  }
+}
