+++ src/main/java/org/folio/rest/converter/PackagesConverter.java
@@ -1,42 +1,22 @@
 package org.folio.rest.converter;
 
-import static org.folio.rest.util.RestConstants.PACKAGES_TYPE;
-import static org.folio.rest.util.RestConstants.PROVIDERS_TYPE;
-import static org.folio.rest.util.RestConstants.RESOURCES_TYPE;
-
 import java.util.EnumMap;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 import org.folio.rest.jaxrs.model.ContentType;
 import org.folio.rest.jaxrs.model.Coverage;
 import org.folio.rest.jaxrs.model.HasManyRelationship;
 import org.folio.rest.jaxrs.model.HasOneRelationship;
 import org.folio.rest.jaxrs.model.MetaDataIncluded;
-import org.folio.rest.jaxrs.model.MetaTotalResults;
-import org.folio.rest.jaxrs.model.Package;
-import org.folio.rest.jaxrs.model.PackageCollection;
-import org.folio.rest.jaxrs.model.PackageCollectionItem;
 import org.folio.rest.jaxrs.model.PackageDataAttributes;
 import org.folio.rest.jaxrs.model.PackagePostRequest;
 import org.folio.rest.jaxrs.model.PackagePutRequest;
 import org.folio.rest.jaxrs.model.PackageRelationship;
-import org.folio.rest.jaxrs.model.Proxy;
-import org.folio.rest.jaxrs.model.RelationshipData;
-import org.folio.rest.jaxrs.model.VisibilityData;
-import org.folio.rest.util.RestConstants;
 import org.folio.rmapi.model.CoverageDates;
-import org.folio.rmapi.model.PackageByIdData;
-import org.folio.rmapi.model.PackageData;
 import org.folio.rmapi.model.PackagePost;
 import org.folio.rmapi.model.PackagePut;
-import org.folio.rmapi.model.Packages;
-import org.folio.rmapi.model.Titles;
 import org.folio.rmapi.model.TokenInfo;
-import org.folio.rmapi.model.VendorById;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -64,101 +44,6 @@ public class PackagesConverter {
     contentTypeToRMAPICode.put(ContentType.PRINT, 5);
     contentTypeToRMAPICode.put(ContentType.UNKNOWN, 6);
     contentTypeToRMAPICode.put(ContentType.ONLINE_REFERENCE, 7);
-  }
-
-  @Autowired
-  private CommonAttributesConverter commonConverter;
-  @Autowired
-  private VendorConverter vendorConverter;
-  @Autowired
-  private ResourcesConverter resourcesConverter;
-
-  public PackageCollection convert(Packages packages) {
-    List<PackageCollectionItem> packageList = packages.getPackagesList().stream()
-      .map(this::convertPackage)
-      .collect(Collectors.toList());
-    return new PackageCollection()
-      .withJsonapi(RestConstants.JSONAPI)
-      .withMeta(new MetaTotalResults().withTotalResults(packages.getTotalResults()))
-      .withData(packageList);
-  }
-
-  public Package convert(PackageByIdData packageByIdData) {
-    return convert(packageByIdData, null, null);
-  }
-
-  public Package convert(PackageByIdData packageByIdData, VendorById vendor, Titles titles) {
-    Package packageData = new Package()
-      .withData(convertPackage(packageByIdData))
-      .withJsonapi(RestConstants.JSONAPI);
-
-    packageData.getData()
-      .withRelationships(createEmptyPackageRelationship())
-      .withType(PACKAGES_TYPE)
-      .getAttributes()
-        .withProxy(convertToProxy(packageByIdData.getProxy()))
-        .withPackageToken(commonConverter.convertToken(packageByIdData.getPackageToken()));
-
-    if (titles != null) {
-      packageData.getData()
-        .withRelationships(new PackageRelationship()
-          .withResources(new HasManyRelationship()
-            .withMeta(new MetaDataIncluded()
-              .withIncluded(true))
-            .withData(convertResourcesRelationship(packageByIdData, titles))));
-
-      packageData
-        .getIncluded()
-          .addAll(resourcesConverter.convertFromRMAPIResourceList(titles).getData());
-    }
-
-    if (vendor != null) {
-      packageData.getIncluded().add(vendorConverter.convertToProvider(vendor).getData());
-      packageData.getData()
-        .getRelationships()
-        .withProvider(new HasOneRelationship()
-          .withData(new RelationshipData()
-            .withId(String.valueOf(vendor.getVendorId()))
-            .withType(PROVIDERS_TYPE)));
-    }
-
-    return packageData;
-  }
-
-  private PackageCollectionItem convertPackage(PackageData packageData) {
-    Integer providerId = packageData.getVendorId();
-    String providerName = packageData.getVendorName();
-    Integer packageId = packageData.getPackageId();
-    return new PackageCollectionItem()
-      .withId(providerId + "-" + packageId)
-      .withType(PACKAGES_TYPE)
-      .withAttributes(new PackageDataAttributes()
-        .withContentType(contentTypes.get(packageData.getContentType().toLowerCase()))
-        .withCustomCoverage(
-          new Coverage()
-            .withBeginCoverage(packageData.getCustomCoverage().getBeginCoverage())
-            .withEndCoverage(packageData.getCustomCoverage().getEndCoverage()))
-        .withIsCustom(packageData.getIsCustom())
-        .withIsSelected(packageData.getIsSelected())
-        .withName(packageData.getPackageName())
-        .withPackageId(packageId)
-        .withPackageType(packageData.getPackageType())
-        .withProviderId(providerId)
-        .withProviderName(providerName)
-        .withSelectedCount(packageData.getSelectedCount())
-        .withTitleCount(packageData.getTitleCount())
-        .withAllowKbToAddTitles(packageData.getAllowEbscoToAddTitles())
-        .withVisibilityData(
-          new VisibilityData().withIsHidden(packageData.getVisibilityData().getIsHidden())
-            .withReason(
-              packageData.getVisibilityData().getReason().equals("Hidden by EP") ? "Set by system"
-                : "")))
-      .withRelationships(createEmptyPackageRelationship());
-  }
-
-  private Proxy convertToProxy(org.folio.rmapi.model.Proxy proxy) {
-    return proxy != null ? new Proxy().withId(proxy.getId())
-      .withInherited(proxy.getInherited()) : null;
   }
 
   public PackagePut convertToRMAPICustomPackagePutRequest(PackagePutRequest request) {
@@ -212,8 +97,8 @@ public class PackagesConverter {
 
     return builder;
   }
-  public PackagePost convertToPackage(PackagePostRequest postPackageBody) {
 
+  public PackagePost convertToPackage(PackagePostRequest postPackageBody) {
     PackagePost.PackagePostBuilder postRequest = PackagePost.builder()
       .contentType(contentTypeToRMAPICode.getOrDefault(postPackageBody.getData().getAttributes().getContentType(), 6))
       .packageName(postPackageBody.getData().getAttributes().getName());
@@ -230,16 +115,7 @@ public class PackagesConverter {
     return postRequest.build();
   }
 
-  private List<RelationshipData> convertResourcesRelationship(PackageByIdData packageByIdData, Titles titles) {
-    return titles.getTitleList().stream()
-      .map(title ->
-        new RelationshipData()
-          .withId(packageByIdData.getVendorId() + "-" + packageByIdData.getPackageId() + "-" + title.getTitleId())
-          .withType(RESOURCES_TYPE))
-      .collect(Collectors.toList());
-  }
-
-  private static PackageRelationship createEmptyPackageRelationship() {
+  public static PackageRelationship createEmptyPackageRelationship() {
     return new PackageRelationship()
       .withProvider(new HasOneRelationship()
         .withMeta(new MetaDataIncluded().withIncluded(false)))
